+++ tests/fixtures.py
@@ -97,3 +97,13 @@ def reduced_cdataset():
         means=[29.3, 28.4, 27.2, 26.0, 25.8, 24.5],
         stdevs=[2.53, 1.9, 2.53, 2.53, 2.21, 1.58]
     )
+
+
+@pytest.fixture
+def ddataset_requires_dose_drop():
+    # a dataset which requires doses to be dropped before a model can be recommended
+    return bmds.DichotomousDataset(
+        doses=[0.0, 10.0, 20.0, 40.0, 80.0, 120.0],
+        ns=[10, 10, 10, 10, 10, 10],
+        incidences=[0, 0, 3, 4, 6, 0],
+    )